Moved checker classes into `check` subpackage.
